requirements.txt
boto3==1.26.142
boto3-stubs[essential]==1.26.142
boto3-stubs-lite[essential]==1.26.142 # pycharm
ffmpeg-python
python-dotenv
requests
mariadb